# lib/api-zod/src/generated/types/contactSubmission.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactSubmission {
  id: number;
  name: string;
  /** @nullable */
  company?: string | null;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  assetClass?: string | null;
  message: string;
  createdAt: Date;
}