# lib/api-zod/src/generated/types/contactSubmissionInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactSubmissionInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  /**
     * @maxLength 200
     * @nullable
     */
  company?: string | null;
  /** @maxLength 320 */
  email: string;
  /**
     * @maxLength 50
     * @nullable
     */
  phone?: string | null;
  /**
     * @maxLength 100
     * @nullable
     */
  assetClass?: string | null;
  /**
     * @minLength 1
     * @maxLength 5000
     */
  message: string;
  /**
     * Honeypot field. Must be left empty by real users.
     * @nullable
     */
  website?: string | null;
}